new flow of listener
